requirements.txt
reflex>=0.5.5
sqlalchemy_utils
opyml
tiktoken
feedparser
feedgen
httpx
dateutils
cityhash
bs4
lxml
lxml_html_clean
python-dateutil
html2text
mistune
newspaper4k
feed2json
deepl
PyDeepLX
anthropic
openai
google-generativeai
easytranslator
translators
langdetect